+++ backend/canisters/community/impl/src/updates/accept_p2p_swap.rs
@@ -47,7 +47,12 @@ async fn accept_p2p_swap(args: Args) -> Response {
                         thread_root_message_index,
                         message_id.into(),
                     ) {
-                        if channel.chat.members.contains(&message.sender) {
+                        if channel
+                            .chat
+                            .members
+                            .get(&message.sender)
+                            .map_or(false, |m| !m.user_type.is_bot())
+                        {
                             let community_id = state.env.canister_id().into();
 
                             state.data.user_event_sync_queue.push(