+++ backend/canisters/group/impl/src/updates/accept_p2p_swap.rs
@@ -46,7 +46,13 @@ async fn accept_p2p_swap(args: Args) -> Response {
                         .events
                         .message_internal(EventIndex::default(), thread_root_message_index, message_id.into())
                 {
-                    if state.data.chat.members.contains(&message.sender) {
+                    if state
+                        .data
+                        .chat
+                        .members
+                        .get(&message.sender)
+                        .map_or(false, |m| !m.user_type.is_bot())
+                    {
                         state.data.user_event_sync_queue.push(
                             message.sender,
                             GroupCanisterEvent::MessageActivity(MessageActivityEvent {